MAINT-7664, MAINT-7681: Make message reflow; add no-wait UI option.
The Tkinter Message widget automatically reflows its contained text to fit the
size of the widget. The trick is to fit the widget properly into its
containing frame. Introduce a place_message() method to set that up.

Make that Message widget display the contents of a StringVar rather than a
constant. Introduce a set_message() method to set content and color
appropriately. In fact, given set_message(), we can instantiate the Message
widget in the InstallerUserMessage constructor instead of repeating the logic
in each display method.

Add optional wait=True parameter to each of the existing display methods. If
you explicitly pass wait=False, the method displays the frame but immediately
returns to its caller. This allows a caller to display, for instance, a
basic_message() but continue processing. set_message() allows updating the
displayed text; an eventual destroy() call removes the status frame again.

But when wait=True either explicitly or by default, binary_choice_message(),
trinary_choice_message() and trinary_choice_link_message() can simply return
the user's selected value to the caller.

Reorder logic in trinary_choice_link_message() so that it's a strict superset
of trinary_choice_message(); this allows the former to simply call the latter.

Honor the progress_bar(interval) parameter.

Fix minor bug in auto_resize().