MAINT-7571: Eliminate ambiguity in get_settings() function.
The get_settings() function accepted a parameter that was either a directory
path or a full file pathname, along with an optional filename in case you
wanted to specify some other file in the passed directory name.

Eliminate the latter parameter, which was never used.

The code checked whether the passed pathname is a file, and decided on that
basis whether to append "user_settings/settings.xml". Unfortunately, when the
parameter was the pathname of a settings file that didn't yet exist, this
resulted in appending "user_settings/settings.xml" to the file pathname,
producing a bogus path.

Unconditionally treat the only remaining parameter as the pathname of the
desired settings file. Change all existing calls accordingly.

Add Application.user_settings_path() to centralize the path logic used for
both get_settings() and write_settings().