MAINT-7639, MAINT-7692: Use 32-bit viewer on Windows 8.1 and 10.
Testing the Alex Ivy viewer on Windows 8.1 with dubious Intel graphics cards
empirically suggests that the 64-bit viewer will not run, whereas the 32-bit
viewer will. So on a system with those graphics cards, fall back to the 32-bit
viewer for both Windows 10 and Windows 8.1.

Since one of the hardest problems in computer science is naming things,
naturally this means we must remove "Windows10" from the names of the relevant
functions and classes.