MAINT-8087: Work around bug in Windows Python 2 subprocess.Popen().
Seems if you pass Windows subprocess.Popen() a non-ASCII command string, it
fails due to encoding errors. Someone on stackoverflow claims that Python 2
uses the Windows CreateProcessA() API, which takes ASCII-only names, but that
Python 3 uses CreateProcessW() and handles this correctly.

Be that as it may, as long as the program filename itself is plain ASCII, it
works to chdir() to its containing directory and then execute just the
filename. Evidently os.chdir() is cleverer about Unicode pathnames.