MAINT-8135: Call CommandLineToArgvW(GetCommandLineW()).
There is a documented special case for CommandLineToArgvW(): if you pass it an
empty string instead of the actual command line, it's supposed to return just
the pathname of the running executable. Since that's exactly what we need to
fix MAINT-8078, our first attempted fix didn't bother with GetCommandLineW()
-- we just passed the empty string to CommandLineToArgvW(), as documented.

BUT on this special code path, CommandLineToArgvW() scans and splits the
command's pathname! So if the path to the current executable contains spaces
(e.g. "c:\Program Files\Something\Something"), then you get back a list
containing [u'C:\\Program', u'Files\\Something\\Something']: the well-known
Windows idiocy concerning pathnames with spaces. If, however, you actually
call GetCommandLineW() and pass *that* string to CommandLineToArgvW(), then
the complete command pathname, spaces and all, is returned in the first entry.

Do that. :-P