MAINT-8304: Engage Tkinter/eventlet interleaving much sooner.
The previous logic assumed we wouldn't need to run Tkinter's mainloop() until
we're ready to listen for clicks on the Dock/Taskbar icon, that is, until
we've actually launched the viewer child process. Unfortunately that means
Tkinter doesn't even start *trying* to display UI until after we've asked it
to hide the VMP's main StatusMessage window.

Before the InstallerUserMessage.py refactoring, we had to ensure that every
change to any UI element was followed by a Tk update() call to force Tkinter
to redraw its UI. These were removed in the recent refactoring, per Tkinter
convention. But we still need to force redrawing somehow. mainloop() is
usually responsible for that.

The solution is to launch the TkGreenthread eventlet greenthread early in VMP
processing, so that every UI change is quickly followed by (re-)entry to
mainloop(). We still don't need to (can't) bind our responses to Dock/Taskbar
icon clicks until we've launched the viewer and have its PID, but we can make
those bind() calls on the existing StatusMessage window -- we don't need to,
and shouldn't, defer window creation to the point where we have the viewer's
PID.