MAINT-8304: More carefully distinguish package names from pathnames.
We used to drive the package-copy logic from a simple list of package names.
That has been changed to a set of pathnames added to the virtualenv's
site-packages directory by pip install. So at copy time, we can skip the
dynamic-import logic used to convert from a plain package name to the pathname
where that package is installed.

However, to cover the case in which one or more packages is already installed
in system Python, we must still ensure that our basic requirements are copied.
That *does* require converting from package names to pathnames.