MAINT-8304: Pass an open pipe to viewer's stdin for EOF detection.
If the viewer hangs a read() on the pipe passed as its stdin, that read() will
block until either we write something or the pipe is closed. Since we have no
intention of either writing or explicitly closing the pipe, it will be
implicitly closed when the VMP terminates. Thus, when the viewer's read()
returns EOF, the viewer can infer VMP termination and enter its own shutdown
processing.

Looking ahead, we could potentially use that pipe to communicate other
information to the viewer than simply, "Gack, I'm dead."