MAINT-8304: Also log any plain print output from anywhere.
In a windowed-app environment such as the VMP, simple print output is normally
discarded. In some Python environments, filling the first sys.stdout buffer
can actually crash the process when it tries to write to an invalid file
handle.

Naturally, resetting sys.stdout and sys.stderr has no effect if no one ever
writes to either of them.

But for debugging external packages that don't have access to vmp_util
.SL_Logging, it's very useful to be able to throw in a print statement and
have it emitted to the log file.