MAINT-8304: Use not-yet-official fork of eventlet package.
Change the definition and uses of BUILD_DEPS and RUNTIME_DEPS to allow pip
installing from a remote URL as well as a simple package name. For eventlet,
specify a URL corresponding to the source of the pull request to merge in
changes essential to running in the PyInstaller frozen runtime environment.

Similarly, although 'enum34' is the name used by pip install, 'enum' is the
name we must import and copy.

On Windows, the only file we really need from the viewer_manager autobuild
package tarball is SL_Launcher.exe. Restructure build-cmd.py so that on
Windows, we don't even copy the other stuff into the stage directory.

Move the logic that deduces what packages we pip installed, and the import
logic to discover where they live, and the loop that copies them, into the
specific Mac/Linux stanza.

Direct PyInstaller to find the SL_Launcher source module (and its
dependencies) in vmp-src, but to place the resulting .exe in stage/VMP.