MAINT-8304: Don't let subprocess_args() hide the viewer window.
On Windows, most of the time, subprocess_args() is used to (set up arguments
to) run a helper child process, a console program such as wmic. Hiding the
associated console window is very appropriate for such cases.

Now that we're using subprocess_args() to (help) run the viewer itself,
though, we must NOT hide the child-process window!