DRTVWR-474: Set SkipBenchmark in user settings file, not cmd line.
On Windows, the Poseidon updater is run before the first viewer run. If it
discovers that it needs to suppress viewer graphics benchmarking, passing a
switch on that first run's command line wouldn't suffice -- the second run
would attempt graphics benchmarking and crash.

Thereafter, it's the viewer that runs the updater. If the viewer attempts
graphics benchmarking when it shouldn't, it will crash before running the
updater.

Therefore, if that first updater run must suppress viewer graphics
benchmarking, it must set it as one of the persistent user settings so it
affects every subsequent viewer run.