DRTVWR-474: Refactor leap.py to defer protocol initialization.
The leap.py module, as adopted from VITA, used a (very) nonstandard approach
to initialization: reading its stdin at module scope, that is, at import time.
In Poseidon, this required deferring the import, which meant jumping through
an increasing number of hoops to avoid referencing (e.g.) exceptions defined
in leap.py before importing it.

Move that logic to module-scope __init__() function, permitting normal
consumer imports. Make leapcomm.ViewerClient.__init__() call leap.__init__().
Fix the other weird workarounds for the deferred leap import.

Also, fix a ViewerClient bug in the rare-but-interesting case in which a
reqid-stamped request is sent with send() and retrieved by a WaitFor
subclass -- rather than using request() as usual. We normally pop the reqid
when checking for a corresponding WaitForReqid instance. But in this case, we
were handing the incoming response to _unsolicited(), where a WaitFor subclass
was looking for that reqid. Since we'd already popped it, it wasn't there!