DRTVWR-474: Reorg optional-update code for possible fast downloads.
For a leap() mode optional update, handle the 'skip' case first. Introduce new
process_optional_update() function for the cases in which we either have, or
have not yet, downloaded the optional update. When we have not, download it
inline on the main thread, but without download UI. If it finishes before the
user clicks Login, treat it the same as a previously-downloaded installer.

When we sync with the viewer, honor the viewer's recommendation as to whether
or not to proceed with update. If login fails due to bad credentials, since
the login-fail popup says nothing about a pending update, it would be
surprising at best to proceed with an install on viewer termination.

Defend each ViewerClient.request() call against ViewerShutdown exceptions. In
some cases, we can interpret the user closing the viewer in response to our
modal popup as permission to proceed with the install.

Defend ViewerClient.shutdown() itself against repeated calls.

When download_update() explicitly downloads a new installer, it already has
its pathname in hand. Make update_manager.download() return that pathname.
Make update_manager.install() and apply_update.apply_update() accept it,
instead of having to dig it out of the download directory.

When we download an installer, capture its pathname. When we simply detect a
previous download, call apply_update.get_filename() to discover the full
pathname of the previously-downloaded installer. Either way, pass the
installer name to update_manager.install().

Make apply_update.get_filename(), which finds a platform-appropriate installer
given the download directory, log both absence and ambiguity.

Make update_manager.download() accept a bool 'ui' parameter instead of
'background'. In practice, 'ui' simply inverts the sense of 'background'. The
difference is in caller expectations and behavior: in a separate updater
process, it's entirely reasonable to call download() on the main thread, even
when the viewer is charging ahead at full speed. In that case we simply want
to suppress download()'s UI. In case of error, make download() unconditionally
raise exception, instead of swallowing it if ui is False.

Introduce updater.download() and install() convenience wrappers for the
respective update_manager functions.