SL-10153: Override passed viewer exe with Application.executable().
NSIS passes the viewer executable to run after precheck(). But if the viewer
is installed into a directory path containing non-ASCII characters, Python
fumbles the non-ASCII characters into ASCII '?'. However, SLVersionChecker can
already determine the intended name of the target viewer. Compare the passed
viewer pathname with the one determined internally, and use the internal one
if they differ (with a log warning).