SL-11216: Post release notes from VVM to viewer, or at least guess.
If the VVM provides a response containing 'more_info', post that to the
viewer's "relnotes" LLEventPump.

If it doesn't, guess a release-notes URL and post that instead.

Also fetch the actual suite of EStartupState enum name strings and look up the
names we care about, instead of embedding a very unobvious requirement to
maintain a couple local symbols matching the EStartupState enum in the viewer
repository.