Find PyInstaller by importing it.
The previous logic jumps through hoops to search for pyinstaller.exe on the
PATH, or, failing that, pyinstaller-script.py, so it can run PyInstaller as a
subprocess. But PyInstaller documents that you can import it and call a run()
entry point -- way simpler and more robust!

Also determine the bit width of the running Python interpreter using struct to
discover the size of pointers, rather than finding '32 bit' in version string.

Also call run(sys.executable, -m, pip, ...) instead of run(pip) to avoid the
possibility of finding a different pip than the one associated with the
running interpreter.