Use PyInstaller on all target systems, not just Windows.
Our macOS strategy to this point has been to bundle up Python packages and
local modules unlikely to be found on system Python, relying on system Python
to run the updater script. But system Python is Python 2. Now that we depend
on Python 3, use PyInstaller to make a complete self-contained executable on
macOS as well as on Windows. This greatly simplifies the build script.