SL-13792: Remove funky Python 2 non-ASCII workarounds.
We don't have to force a non-empty bogus TIX_LIBRARY environment variable any
more because Python 3 tkinter should Just Work.

We don't have to cross-check the viewer pathname passed into the precheck
function against what we believe to be the viewer's pathname.

When launching the viewer shortcut, we no longer have to cd to the install
directory and then launch .\SecondLife.lnk : we can just launch the correct
pathname. In fact we can get rid of the whole PopenRunner(cwd=) feature, which
was introduced for that specific purpose.

We no longer have to create and execute a nextviewer.bat Windows batch file to
run the installer. That actually dates back to our flirtation with Windows
user installs, which we decided against supporting.

We no longer need a Runner.Popen() function. That was solely to allow running
an executable with a non-ASCII pathname by changing to the containing
directory and then executing the unqualified filename.

We no longer need the util.ufile() or udir() workarounds, but make the
functions trivial instead of removing them entirely.

We no longer need to query the Windows API directly to get the name of the
running executable. Remove the whole util.Application.get_executable_name()
function.

We can use pathlib.Path to present the individual components of a pathname.

We can get the Windows Application Data path from $APPDATA rather than having
to query the Windows API directly for it.