SL-13792: Fix unit tests on Mac.
It seems weirdly as though maybe self-tests haven't been run on Mac in a
while? Lots of errors of the form "How the bleep did this work before?"

Also update to current PyInstaller: locking it to version 3.5 is now
counterproductive.

Also update to nose2: it seems Python 3.10 breaks nose (ouch!).