SL-16842, SL-14300: distutils copy_tree() => shutil.copytree().
This should get us out of the business of trying to predict which distutils
helper modules will be lazily loaded by distutils.dir_util.copy_tree(), the
cause of SL-16842.

Moreover, shutil.copytree() supports a per-file callback, which gives us the
chance to drive the updater's existing progress bar to monitor the copy
operation.

Allow a specific failure message for InstallerUserMessage.intercept_close() so
that if the copy operation fails for any reason, we can clarify in the log
that that's what we were trying to do.